list of lists being scraped, lists being scraped, plex playlists being generated succesfully
