+++ lists/list_names.json
@@ -1,0 +1,54 @@
+[
+  {
+    "name": "Ambient",
+    "url": "https://letterboxd.com/larswan/list/ambient/"
+  },
+  {
+    "name": "WOULD Rewatch \ud83d\udc69\u200d\ud83c\udfeb\ud83e\udee1",
+    "url": "https://letterboxd.com/larswan/list/should-rewatch/"
+  },
+  {
+    "name": "Perfect for what they are",
+    "url": "https://letterboxd.com/larswan/list/perfect-for-what-they-are/"
+  },
+  {
+    "name": "Comfort \ud83c\udf5c \ud83d\ude1a \ud83d\udc93",
+    "url": "https://letterboxd.com/larswan/list/comfort-movies/"
+  },
+  {
+    "name": "Winter \u2744\ufe0f \ud83e\udd76",
+    "url": "https://letterboxd.com/larswan/list/winter/"
+  },
+  {
+    "name": "Britflix",
+    "url": "https://letterboxd.com/larswan/list/britflix/"
+  },
+  {
+    "name": "Old Asian",
+    "url": "https://letterboxd.com/larswan/list/old-asian/"
+  },
+  {
+    "name": "Favorites \u2b50\ufe0f\ud83c\uddfa\ud83c\uddf8",
+    "url": "https://letterboxd.com/larswan/list/favorites/"
+  },
+  {
+    "name": "Casual \ud83e\udd37\u200d\u2640\ufe0f\u2757\ufe0f\ud83e\udd37",
+    "url": "https://letterboxd.com/larswan/list/casual/"
+  },
+  {
+    "name": "Serious movie time \ud83c\udf7f",
+    "url": "https://letterboxd.com/larswan/list/serious-movie-time/"
+  },
+  {
+    "name": "Reset/Inspire",
+    "url": "https://letterboxd.com/larswan/list/reset-inspire/"
+  },
+  {
+    "name": "Nighttime \ud83c\udf19 \ud83c\udf03",
+    "url": "https://letterboxd.com/larswan/list/late-night/"
+  },
+  {
+    "name": "Interested \ud83e\udd78",
+    "url": "https://letterboxd.com/larswan/list/interested/"
+  }
+]
